+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-# 1.0.0
+# Changelog
+
+## 1.1.0
+
+- Code is split into multiple modules
+- More command-line options (port and timeout duration)
+- Some code documentation
+- Better(?) input validation
+- Better error handling (using `color-eyre`)
+
+## 1.0.0
 
 - First version